move HomeOnboarding to a separate component
